poc/requirements.txt
langgraph>=0.2.30
sqlglot>=25.6.0
SQLAlchemy>=2.0.30
pydantic>=2.8.2
python-dotenv>=1.0.1
openai>=1.40.0
pandas>=2.2.2
tabulate>=0.9.0
